OCR/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pytesseract==0.3.10
Pillow==10.0.1
pdf2image==1.16.3
google-generativeai==0.3.2
pandas==2.1.1
python-dotenv==1.0.0
